feat(feedback): Flush replays when feedback form opens (#10567)
Flush replay when the feedback form is first opened instead of at submit
time

We are making this change because we have noticed a lot of feedback
replays only consist of the user submitting the feedback and not what
they did prior to submitting feedback. This may result in false
positives if users open but do not submit feedback, but this should make
replays from feedback more useful.
